Renaming packages to diagnostics
